Optim: avoid some iterations while avoiding some iterations
