+++ normalized-cache-incubating/src/commonMain/kotlin/com/apollographql/cache/normalized/internal/CacheBatchReader.kt
@@ -57,6 +57,11 @@ internal class CacheBatchReader(
    */
   private var isStale = false
 
+  /**
+   * True if at least one of the resolved fields is an Error, or if a cache miss happened
+   */
+  private var hasErrors = false
+
   private val pendingReferences = mutableListOf<PendingReference>()
 
   private class CollectState(val variables: Executable.Variables) {
@@ -104,8 +109,13 @@ internal class CacheBatchReader(
     )
 
     while (pendingReferences.isNotEmpty()) {
-      val records = cache.loadRecords(pendingReferences.map { it.key }, cacheHeaders).associateBy { it.key }
-
+      val records: Map<CacheKey, Record> = cache.loadRecords(pendingReferences.map { it.key }, cacheHeaders)
+          .also {
+            if (!hasErrors) {
+              hasErrors = it.any { it.values.any { it.hasErrors() } }
+            }
+          }
+          .associateBy { it.key }
       val copy = pendingReferences.toList()
       pendingReferences.clear()
       copy.forEach { pendingReference ->
@@ -118,6 +128,7 @@ internal class CacheBatchReader(
             if (returnPartialResponses) {
               data[pendingReference.path] =
                 cacheMissError(CacheMissException(key = pendingReference.key.keyToString(), fieldName = null, stale = false), path = pendingReference.path)
+              hasErrors = true
               return@forEach
             } else {
               throw CacheMissException(pendingReference.key.keyToString())
@@ -149,6 +160,7 @@ internal class CacheBatchReader(
           } catch (e: CacheMissException) {
             if (e.stale) isStale = true
             if (returnPartialResponses) {
+              hasErrors = true
               cacheMissError(e, pendingReference.path + it.responseName)
             } else {
               throw e
@@ -163,7 +175,11 @@ internal class CacheBatchReader(
       }
     }
 
-    return CacheBatchReaderData(data, CacheHeaders.Builder().apply { if (isStale) addHeader(ApolloCacheHeaders.STALE, "true") }.build())
+    return CacheBatchReaderData(
+        data = data,
+        cacheHeaders = CacheHeaders.Builder().apply { if (isStale) addHeader(ApolloCacheHeaders.STALE, "true") }.build(),
+        hasErrors = hasErrors,
+    )
   }
 
   private fun Any?.unwrap(): Any? {
@@ -234,6 +250,7 @@ internal class CacheBatchReader(
           } catch (e: CacheMissException) {
             if (e.stale) isStale = true
             if (returnPartialResponses) {
+              hasErrors = true
               cacheMissError(e, path + it.responseName)
             } else {
               throw e
@@ -248,6 +265,7 @@ internal class CacheBatchReader(
   internal class CacheBatchReaderData(
       private val data: Map<List<Any>, Any>,
       val cacheHeaders: CacheHeaders,
+      val hasErrors: Boolean,
   ) {
     @Suppress("UNCHECKED_CAST")
     internal fun toMap(withErrors: Boolean = true): DataWithErrors {
@@ -303,5 +321,21 @@ internal class CacheBatchReader(
         .path(path = path)
         .cacheMissException(exception)
         .build()
+  }
+
+  @Suppress("UNCHECKED_CAST")
+  internal fun Any?.hasErrors(): Boolean {
+    val queue = ArrayDeque<Any?>()
+    queue.add(this)
+    while (queue.isNotEmpty()) {
+      val current = queue.removeFirst()
+      when (current) {
+        is Error -> return true
+        is List<*> -> queue.addAll(current)
+        // Embedded fields can be represented as Maps
+        is Map<*, *> -> queue.addAll(current.values)
+      }
+    }
+    return false
   }
 }
